backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.10.0
pydantic==2.9.0
httpx==0.27.0
textblob==0.18.0
